add correction

+++ src/Service/Calculator.php
@@ -4,7 +4,7 @@ namespace App\Service;
 
 class Calculator
 {
-    public function add(int $a, int $b): int
+    public function add(float $a, float $b): float
     {
         return $a + $b;
     }
